packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tunalab-core"
version = "0.1.0"
description = "Core runtime utilities for TunaLab"
dependencies = [
    "torch",
    "numpy",
    "pyyaml",
    "tqdm",
]

[tool.setuptools.packages.find]
where = ["src"]